requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
aiohttp==3.9.1
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
websockets==12.0
schedule==1.2.0
python-multipart==0.0.6
jinja2==3.1.2
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
httpx==0.25.2 